Two data

+++ modules/Employe/Http/Controllers/EmployeController.php
@@ -21,7 +21,7 @@ class EmployeController extends Controller
     {
         $req = request();
         if ($req->ajax()) {
-            return DataTables::of(Employe::all())
+            return DataTables::of(Employe::where('noreg', '!=' , null)->get())
                 ->editColumn('nokta', function($row) {
                     return $row->nokta . ' - ' . $row->noregu;
                 })
@@ -38,6 +38,29 @@ class EmployeController extends Controller
                 ->make(true);
         }
         return view('employe::index');
+    }
+
+    public function index1()
+    {
+        $req = request();
+        if ($req->ajax()) {
+            return DataTables::of(Employe::where('noreg', null)->get())
+                ->editColumn('nokta', function($row) {
+                    return $row->nokta . ' - ' . $row->noregu;
+                })
+                ->addColumn('action', function($row) {
+                    $btn = '<div class="btn-group" role="group">';
+
+                    $btn .= '<a href="'.route('employe::read', ['id' => $row->id]).'" class="btn btn-sm btn-info"><i class="c-icon material-icons">drafts</i></a>';
+                    $btn .= '<a href="'.route('employe::edit', ['id' => $row->id]).'" class="btn btn-sm btn-warning"><i class="c-icon material-icons">print</i></a>';
+
+                    $btn .= '</div>';
+
+                    return $btn;
+                })
+                ->make(true);
+        }
+        return view('employe::tkbm');
     }
 
     /**
@@ -127,16 +150,27 @@ class EmployeController extends Controller
             $request->session()->put('form_steps', Crypt::encryptString(json_encode($data)));
             return redirect(route('employe::edit', ['id' => $id]));
         } else if ($request->form_step === '2') {
+            if ($request->file('photo-default')) {
+                $request->validate([
+                    'photo-default' => 'image'
+                ]);
+            }
+
             $form_steps = $this->updateFormSteps(true);
             $form_steps['form_step'] = 3;
-            $form_steps['photo'] = $request->photo;
+            if (empty($request->photo)) {
+                if ($request->file('photo'))
+                    $form_steps['photo'] = $request->file('photo-default')->store('public/berkas');
+            } else {
+                $form_steps['photo'] = $request->photo;
+            }
             $this->updateFormSteps(false, $form_steps);
             return redirect(route('employe::edit', ['id' => $id]));
         } else {
             $request->validate([
-                'photo_ktp' => 'required|image',
-                'photo_kk' => 'required|image',
-                'photo_kta' => 'required|image',
+                'photo_ktp' => 'required',
+                'photo_kk' => 'required',
+                'photo_kta' => 'required',
                 'ijazah' => 'image',
                 'sertifikat' => 'image',
             ]);
@@ -146,8 +180,9 @@ class EmployeController extends Controller
             $form_steps['serikat_id'] = $form_steps['serikat'];
             $q1 = $form_steps['kursus'];
             $q2 = $form_steps['sertifikasi'];
-            $photo = base64_decode($form_steps['photo']);
-            $form_steps['tgllahir'] = date('Y-m-d', strtotime($form_steps['tgllahir']));
+            // $photo = base64_decode($form_steps['photo']);
+            if (!empty($form_steps['tgllahir']))
+                $form_steps['tgllahir'] = date('Y-m-d', strtotime($form_steps['tgllahir']));
             unset(
                 $form_steps['_token'],
                 $form_steps['form_step'],
@@ -156,17 +191,15 @@ class EmployeController extends Controller
                 $form_steps['sertifikasi'],
                 $form_steps['serikat']
             );
-            // try {
-            //     Storage::disk('public')->delete('berkas/' . $form_steps['nama'] . '-' . $form_steps['nokta'] . '.jpeg');
-            // } catch (Exception $e) {}
-            // Storage::disk('public')->put('berkas/' . $form_steps['nama'] . '-' . $form_steps['nokta'] . '.jpeg', $photo);
-            // $form_steps['photo'] = 'berkas/' . $form_steps['nama'] . '-' . $form_steps['nokta'] . '.jpeg';
-            $filename = $form_steps['nama'] . '-' . $form_steps['nokta'] . '-' . $form_steps['noregu'] . '.jpeg';
-            if (!is_dir(storage_path('app/public/berkas'))) {
-                mkdir(storage_path('app/public/berkas'));
+            if (!file_exists(storage_path('app/' . $form_steps['photo']))) {
+                $photo = base64_decode($form_steps['photo']);
+                $filename = $form_steps['nama'] . '-' . $form_steps['nokta'] . '-' . $form_steps['noregu'] . '.jpeg';
+                if (!is_dir(storage_path('app/public/berkas'))) {
+                    mkdir(storage_path('app/public/berkas'));
+                }
+                file_put_contents(storage_path('app/public/berkas/' . $filename), $photo);
+                $form_steps['photo'] = 'berkas/' . $filename;
             }
-            file_put_contents(storage_path('app/public/berkas/' . $filename), $photo);
-            $form_steps['photo'] = 'berkas/' . $filename;
             $data = [
                 'photo_ktp',
                 'photo_kk',
@@ -178,13 +211,15 @@ class EmployeController extends Controller
             foreach ($data as $key) {
                 $file = $request->file($key);
                 if ($file) {
-                    $path = $file->store('berkas/' . $form_steps['nama'] . '-' . $form_steps['nokta']);
+                    $path = $file->store('berkas/' . $form_steps['nama'] . '-' . $form_steps['nokta'] . '-' . $form_steps['noregu']);
                     $files[$key] = $path;
                 }
             }
             $d = Employe::find($id);
             if (empty($d->noreg)) {
                 $form_steps['noreg'] = $this->getNoRegistrasi();
+                $form_steps['tgl_pembuatan'] = date('Y-m-d');
+                $form_steps['user_id'] = auth()->user()->id;
             }
             $c1 = $d->update($form_steps);
             $files['employe_id'] = $d->id;
@@ -218,14 +253,15 @@ class EmployeController extends Controller
         if ($check) {
             $j = str_replace($prefix, '', $check->noreg);
             $n = (int) $j;
-            if (strlen($n) === 1) {
-                $code = $prefix . '000' . ($n + 1);
-            } else if (strlen($n) === 2) {
-                $code = $prefix . '00' . ($n + 1);
-            } else if (strlen($n) === 3) {
-                $code = $prefix . '0' . ($n + 1);
+            $ni = $n + 1;
+            if (strlen($ni) === 1) {
+                $code = $prefix . '000' . $ni;
+            } else if (strlen($ni) === 2) {
+                $code = $prefix . '00' . $ni;
+            } else if (strlen($ni) === 3) {
+                $code = $prefix . '0' . $ni;
             } else {
-                $code = $prefix . ($n + 1);
+                $code = $prefix . $ni;
             }
             // Generate angka baru
             $code = $code;
@@ -238,7 +274,10 @@ class EmployeController extends Controller
     public function check(Request $request) {
         $nokta = $request->get('nokta');
         $noregu = $request->get('noregu');
-        $data = Employe::where('nokta', $nokta)->whereAnd('noregu', $noregu)->first();
+        $data = Employe::where([
+            ['nokta', $nokta],
+            ['noregu', $noregu],
+        ])->first();
         $status = 0;
         $id = null;
         if ($data) {
@@ -265,5 +304,9 @@ class EmployeController extends Controller
             $randomString .= $characters[rand(0, $charactersLength - 1)];
         }
         return $randomString;
+    }
+
+    public function testData() {
+        echo $this->getNoRegistrasi();
     }
 }
